+++ packages/slice-machine/src/modules/selectedCustomType/types.ts
@@ -4,12 +4,18 @@ import {
   TabField,
 } from "@slicemachine/core/build/src/models/CustomType";
 
+export enum CustomTypeStatus {
+  New = "NEW_CT",
+  Modified = "MODIFIED",
+  Synced = "SYNCED",
+}
+
 export type PoolOfFields = ReadonlyArray<{ key: string; value: TabField }>;
 
 export type SelectedCustomTypeStoreType = {
   model: CustomTypeSM;
   initialModel: CustomTypeSM;
-  remoteModel: CustomTypeSM | null | undefined;
+  remoteModel: CustomTypeSM | undefined;
   mockConfig: CustomTypeMockConfig;
   initialMockConfig: CustomTypeMockConfig;
 } | null;
